requirements.txt
fastapi
uvicorn
langchain
openai
load-dotenv
faiss-cpu
pypdf
sentence-transformers
pymupdf
groq
langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface
reportlab
streamlit
